Convert the script to use a CLI
Using the argparse library we can accept input arguments and no longer
rely on updating the piazza code within the script.

I also added the typing module to make some things a big clearer.

I also added some basic error checking for the FileIO. There might be
more error checking that needs to be done such as any problems with the
arguments. I think the argparse library handles these problems though
so there might not be anything left.

For future improvements we might want to format the json better or even
parse the json and discard the unnecessary details.

Not sure if the verbose mode was worth implementing for such a simple
script.