Switch to outputting .json instead of .txt
First, I never set up text formatting for the output because I'm lazy.
Next, I was writing python dictionaries as strings to the .txt files
and mistook the result for JSON.

I've now let go of the idea of formating the output because I don't
care anymore. I did update the script so it converts the python
dictionary to a true JSON string before writing to file.

I want to add media support in the future so I can archive images from
Piazza posts as well.